Limit the size of manifests/signatures sync/upload
Adds new settings to limit the size of manifests and signatures as a safeguard
to avoid DDoS attack during sync and upload operations.
To also prevent this during image upload, this commit configures a
`client_max_body_size` for manifests and signatures Nginx endpoints.
Modify the blob upload to read the layers in chunks.

closes: #532